% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleXchr.R
\name{doubleXchr}
\alias{doubleXchr}
\title{double the values of X chr (for human male samples)}
\usage{
doubleXchr(rs_assay, chr = "chrX")
}
\arguments{
\item{rs_assay}{a Repli-seq assay (data.frame) loaded with readRS() and formatted as chr,start,stop,S1,...,Sx (or S0)}

\item{chr}{level of chr to double the value (default is set to "chrX")}
}
\value{
a Repli-seq assay doubling values of chrX windows (data.frame) and formatted as chr,start,stop,S1,...,Sx
}
\description{
Calculate Repli-seq assay count matrices after doubling the values of chrX windows
}
