% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNoise.R
\name{removeNoise}
\alias{removeNoise}
\title{remove Noise}
\usage{
removeNoise(rs_assay, rs_control, noise_ratios)
}
\arguments{
\item{rs_assay}{a Repli-seq assay (data.frame) loaded with readRS() and formatted as chr,start,stop,S1,...,Sx}

\item{rs_control}{a Repli-seq control assay (data.frame) loaded with readRS() and formatted as chr,start,stop,S0}

\item{noise_ratios}{an array with the ratios in the order of rs_assay fractions calculated with calculateNoiseRatios()}
}
\value{
a Repli-seq assay after removing noise (data.frame) and formatted as chr,start,stop,S1,...,Sx
}
\description{
Calculate Repli-seq assay count matrices after substracting noise
}
