% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repliseqCorrPlot.R
\name{repliseqCorrPlot}
\alias{repliseqCorrPlot}
\title{Calculate and plot correlation plots.}
\usage{
repliseqCorrPlot(rs_assay, corr_method = "pearson", ordering = "original")
}
\arguments{
\item{rs_assay}{a Repli-seq assay (data.frame) loaded with readRS() and formatted as chr,start,stop,S1,...,Sx}

\item{corr_method}{character in either "pearson","spearman" or "kendall" to use as a parameter for cor.test function}

\item{ordering}{a character in either "original","AOE", "FPC", "hclust", "alphabet" to reorder the compared variables (to fill the corrplot() order parameter )}
}
\value{
a mixed correlation plot
}
\description{
Calculate and draw a correlation plot from Repli-seq assay (data.frame) see http://www.sthda.com/french/wiki/visualiser-une-matrice-de-correlation-par-un-correlogramme
}
