% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repliseqDensityPlot.R
\name{repliseqDensityPlot}
\alias{repliseqDensityPlot}
\title{Calculate and plot density plots.}
\usage{
repliseqDensityPlot(rs_assay, grided = F, ncols = 3)
}
\arguments{
\item{rs_assay}{a Repli-seq assay (data.frame) loaded with readRS() and formatted as chr,start,stop,S1,...,Sx}

\item{grided}{boolean (default = F) if True then each fraction s displayed on a different panel, otherwise all the fractions are colored in a single panel}

\item{ncols}{(default = 3) amount of columns when grided == T (see ggplot2 facet_wrap ncol parameter)}
}
\value{
a ggplot object displaying density for the provided repliseq assay
}
\description{
Calculate and draw a density plot from Repli-seq assay (data.frame)
}
