% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeBedgraph.R
\name{writeBedgraph}
\alias{writeBedgraph}
\title{Write Bedgraph files of Repli-seq assays}
\usage{
writeBedgraph(rs_assay, path_file, sample_name, system_separator = "/")
}
\arguments{
\item{rs_assay}{a Repli-seq assay (data.frame) loaded with readRS() and formatted as chr,start,stop,S1,...,Sx (or S0)}

\item{path_file}{a path to the}

\item{sample_name}{a string for the file names}

\item{system_separator}{default is "/" as dedicated for linux file system}
}
\description{
writes one bedgraph file per fraction in the provided Repli-seq assay
}
