% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeBigWig.R
\name{writeBigwig}
\alias{writeBigwig}
\title{Write BigWig files of Repli-seq assays}
\usage{
writeBigwig(
  rs_assay,
  path_file,
  sample_name,
  chromsizes,
  wiggle_start,
  wiggle_step,
  wiggle_span,
  system_separator = "/"
)
}
\arguments{
\item{rs_assay}{a Repli-seq assay (data.frame) loaded with readRS() and formatted as chr,start,stop,S1,...,Sx (or S0)}

\item{path_file}{a path to the files to write}

\item{sample_name}{a string for the file names}

\item{chromsizes}{path to the chromsizes file to pass to WigToBigWig as argument}

\item{wiggle_start}{a numeric value for wiggle header}

\item{wiggle_step}{a numeric value for wiggle header}

\item{wiggle_span}{a numeric value for wiggle header}

\item{system_separator}{default is "/" as dedicated for linux file system}
}
\description{
writes one BigWig file per fraction in the provided Repli-seq assay
}
