% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateNoiseRatios.R
\name{calculateNoiseRatios}
\alias{calculateNoiseRatios}
\title{Calculate noise ratios}
\usage{
calculateNoiseRatios(rs_assay, rs_control)
}
\arguments{
\item{rs_assay}{a Repli-seq assay (data.frame) loaded with readRS() and formatted as chr,start,stop,S1,...,Sx}

\item{rs_control}{a Repli-seq control assay (data.frame) loaded with readRS() and formatted as chr,start,stop,S0}
}
\value{
an array with the ratios in the order of rs_assay fractions
}
\description{
Calculate noise ratios of a Repli-seq assay versus a Repli-seq control
}
