% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateS50.R
\name{calculateS50}
\alias{calculateS50}
\title{Calculate S50.}
\usage{
calculateS50(rs_assay)
}
\arguments{
\item{rs_assay}{a Repli-seq assay (data.frame) loaded with readRS() and formatted as chr,start,stop,S1,...,Sx}
}
\value{
a dataframe composed of genomic coordinates plus S50 : chr,start,stop,S50
}
\description{
Calculate replication timing S50 values from Repli-seq assay (data.frame)
}
