% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateURI.R
\name{calculateURI}
\alias{calculateURI}
\title{calculate URI}
\usage{
calculateURI(rs_x, rs_y)
}
\arguments{
\item{rs_x}{a Repli-seq assay (data.frame) loaded with readRS() and formatted as chr,start,stop,S1,...,Sx}

\item{rs_y}{a Repli-seq assay (data.frame) loaded with readRS() and formatted as chr,start,stop,S1,...,Sx}
}
\value{
a dataframe composed of genomic coordinates plus sum_x, sum_y, mean_xy and URI : chr,start,stop,sum_x,sum_y,mean_xy,URI
}
\description{
Calculate the under replication index of two compared Repli-seq assays
}
