% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRS.R
\name{readRS}
\alias{readRS}
\title{Read Repli-seq assay data (bedgraph format).}
\usage{
readRS(paths_data, fractions)
}
\arguments{
\item{paths_data}{list of paths to repliseq fractions files}

\item{fractions}{list of fractions names}
}
\value{
a dataframe composed of genomic coordinates plus all the specified fractions as for example : chr,start,stop,S1,S2,S3,S4,S5,S6
}
\description{
Bedgraphs with a one line header and with similar organization (same length and sorted)
}
